Formalised clock-rate multiplication within disk drives, discovered that the stepper didn't have ideal behaviour for my timed event loop and hence nailed down the semantics a ilttle more.
(obiter: the 1540 now appears to discern the correct sequence of bits. Framing is off in my test printfs but that's neither here nor there).

+++ Machines/Electron/Electron.cpp
@@ -1008,7 +1008,7 @@ inline void Tape::run_for_cycles(unsigned int number_of_cycles)
 		{
 			if(_is_running)
 			{
-				TapePlayer::run_for_cycles(number_of_cycles);
+				TapePlayer::run_for_cycles((int)number_of_cycles);
 			}
 		}
 		else
